+++ Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 [dependencies]
 anyhow = "1.0.82" 
+clap = { version = "4.5.4", features = ["derive"]}
 comat = "0.1.3"
 csv = { version = "1.3.0", default-features = false }
 flate2 = { version = "1.0.30"}
